stbt.Display.__init__: Move member variables to the top of the function
`Display.on_new_buffer` was failing with:

    AttributeError: Display instance has no attribute 'lock'

Signal handlers (such as `Display.on_new_buffer`) are called
synchronously from the thread of the GStreamer element in question (in
this case, from the `queue` thread). So these signal handlers can be
entered even before we start running the GLib main loop (which is only
needed for receiving GStreamer bus *messages*).