stbt.Display.frames: Display the last frame processed
When `frames` is called from `wait_for_match`, this is usually the frame
that contains the match.

(Once a match is found, `wait_for_match` stops iterating over the
`frames` generator, so the code after the `yield` wasn't being run for
the frame with the match. The `finally` block presumably gets run when
the generator object is destroyed -- in the CPython implementation this
will be when `wait_for_match` returns.)