stbt.Display.teardown: Make debug message less scary-looking
"GLib mainloop is not alive" was the expected, correct behaviour but the
"not alive" part might suggest to someone unfamiliar with stbt internals
that there was a problem. This debug message is intended for debugging
(after the fact) test runs where the recorded video (if the user
specified "stbt run --save-video") is corrupted, possibly because the
GStreamer EOS message wasn't received and thus the GLib mainloop hadn't
terminated.