stbt-control: Raise custom exception if the terminal is too small
This is a follow-up to ac484dd0. That correction printed an error
message if the terminal was too small but didn't raise an exception. The
problem with that solution is that if the terminal is even too small to
print the error message then the tool quits with the unhelpful
`curses.error` exception.

Now the tool raises an exception every time the keymap doesn't fit the
terminal window, but with a helpful exception message.