stbt.argparser: `verbose` value not read from config file
`_debug_level` is supposed to get the value of `verbose` parameter from
the stbt config file if none of `-v` or `--verbose` command line arguments
are specified.

This parameter didn't have any effect because the `default=...` argument
of `parser.add_argument` only updates the internal dictionary of
ArgumentParser but doesn't set `_debug_level`; and `IncreaseDebugLevel`
is only called if ArgumentParser reads any of `-v` or `--verbose`
command line arguments.

This correction sets `_debug_level` to the value of `verbose` and then
overwrites it when IncreaseDebugLevel is called.