stbt record: Simplify logic in `record` function
This is mostly a refactoring, though there is one user-visible change:
The numbering of the saved images now starts from 1 instead of 0. This
isn't significant though: it won't affect existing scripts, and new
scripts will work too.

The motivation for this, in addition to making the code clearer and
shorter, is to allow the next commit (which see).

Why do we need two nested `try` statements? We want to take one final
screenshot on `KeyboardInterrupt` (when the user presses Control-C), or
if the `remote_input` iteration ends (when using
`--control-recorder=file:///dev/stdin` and the user presses Control-D).
But we don't want the `except KeyboardInterrupt` block to take another
screenshot if the user pressed Control-C during the
`write_wait_for_match` triggered from `remote_input` iteration ending.
I'll be able to eliminate the outer `try` in the next commit.