stbt record: Avoid explicit `del`
Commit 47c56330 added the line `del remote_input` so that the destructor
of the `remote_input` iterator (i.e. `stbt-control.main_loop`) is run
(including the `curses` cleanup in the `terminal` context manager of
`stbt-control`, which clears the screen) *before* any unhandled
exception backtraces are printed to the terminal.

This current commit avoids the explicit `del` by not holding any
references to the `stbt-control.main_loop` iterator; if an exception is
raised in `record` (e.g. from `stbt.get_frame` or `stbt.press`) then the
`stbt-control.main_loop` iterator is destroyed as soon as the exception
leaves the `record` stack frame. (This of course relies on CPython's
deterministic destruction, but so did the previous solution with the
explicit `del`; `del` only decrements a reference count.)

Specifically, `remote_input` was a local variable inside the `record`
function. The raised exception contains a reference to each frame on the
stack, and the frame for the `record` function contained a reference to
`remote_input`, so `remote_input` wasn't destroyed until the exception
object itself was destroyed.