extra/runner/run: Show "-t" tag in generated report
If you are running multiple instances of `run` in the same parent
directory (for example if you have multiple systems-under-test connected
to the same PC), or if you intend to merge the testrun directories from
several different PCs into a single report, `run -t "my tag"` adds the
given tag to each testrun directory name so that if two tests happen
to be started at the same second, the directory names won't conflict.

Now, that tag is also shown in the generated `index.html` report. For
example, you can set the tag to the hostname of the system-under-test so
that you can easily tell apart the different systems you were testing.

This introduces a generic mechanism for adding extra columns to the
report: Each line in the file <testrun directory>/extra-columns is
treated as a tab-separated (column name, column value) pair. This will
allow me to add extra data without having to modify `report.py` and
`templates/*.html`.