test_wait_for_match_match_method_param_affects_first_pass: Fix false negative
The first `stbt-run`'s exit status was being ignored, which masked an
error in the call to `wait_for_match`. Furthermore, I had to disable the
second pass (`confirm_method="none"`) to get an observable difference
between the two first-pass methods, because even when the first pass
was giving an incorrect match, the second pass was correctly detecting
that it shouldn't match. This test wants to prove that with a different
first-pass method you can get an incorrect "match" verdict.