extra/runner: Show shorter relative path to test script in report
The table in the report is already short on horizontal space, so we want
the test name to be as short as possible, but still unambiguous. I
assume that the test script is in a git repository, and I use the path
of the script relative to the repository root. We can support other
source control systems as the need arises.

I could have used `git rev-parse --show-toplevel` instead of
`--git-dir`, but that requires git v1.7.0 (whereas `--git-dir` exists
since v1.0).