extra/runner/run: Fix `killtree` on Linux
Commit 2e0e1097 changed `killtree` so that it worked on BSD/OSX, but
broke it on Linux. With Linux's GNU `ps`, the output of
`ps -o ppid=,pid=` is:

    ,pid=
     1921
     1922

instead of the desired:

     1921  1922
     1922  2282

`ps -o ppid= -o pid=` works with both GNU and BSD `ps`.

(`-o` selects which columns to display, and setting the column header to
blank (`<column>=<nothing>`) stops the header line being printed.)