extra/runner: Don't generate the html report twice
We need to run `report` before `recover`, because `recover` needs
the `failure-reason` generated by `report`.

And we need to run `report` again *after* `recover`, because the report
should contain any logs or `extra-columns` information generated by
`recover`.

I don't want to split the failure classification and html generation
into two separate scripts, because it is convenient for the user to
re-classify and re-generate the report for previous test runs by running
a single script (for example `report path/to/my-test-session/2013-*`).

The `--classify-only` and `--html-only` flags introduced in this commit
are undocumented, internal flags. The user doesn't need to care about
these; they are only to make `run` slightly more efficient.