extra/runner: selftests: Increase timeout
To avoid false failures when running the tests in parallel. The error
was:

    test_runner_sigint_twice... FAIL
    error: Didn't find 'test.py...' after 1 second

I've also incleased the sleep in `test_runner_sigint_once` to match the
maximum timeout of 2 seconds for each of the `expect_runner_to_say`
calls. And I've increased the sleep in `test_runner_sigint_twice` to an
even larger number to be safe, but in this case it won't increase the
test run-time because the runner should exit after the second `kill`.