stbt.Display: Fix `restart_source` with the Hauppauge HDPVR
This fixes a regression introduced in 587b70e6. For details on the
HDPVR's behaviour and the desired `restart_source` functionality see
b6135eca.

Without this fix, momentary input video loss with the Hauppauge HDPVR
would result in an infinite† loop of:

    Attempting to recover from video loss: Stopping source pipeline and waiting 5s...
    Restarting source pipeline...
    Restarted source pipeline
    Attempting to recover from video loss: Stopping source pipeline and waiting 5s...

† Until the current wait_for_match's `timeout_secs` timeout, whereupon
the test script would fail.

The only reliable way we have found of recovering the HDPVR after
loss of input video is to tear down the v4l2src's entire pipeline
and create it again.