stbt run: `script` positional argument cannot be specified in config file
It never made sense for the script name to default to a value fetched
from stbt.conf, because unlike the source video pipeline and the remote
control mechanism, you'll probably want to run a different test script
each time you invoke `stbt run`.

This change is necessary for the next commit, so that the script name
is not ambiguous with arguments passed to the script.