extra/runner: Report preserves newlines in "failure reason"
The "failure reason" can be generated by the user, so it can contain
multiple lines. Now "failure reason" is displayed just like the "extra
columns" information, which is also generated by the user.

Note that in the summary report (templates/index.html) the "failure
reason" is displayed without preserving newlines, but it is also
truncated to 30 characters so there's no need to preserve newlines as
the multi-line information won't all be visible.