extra/runner: Web server to allow interactively editing the failure-reason
If you run `server 0.0.0.0:5000` from the directory containing the
testrun results, you can then view the html reports at
`http://hostname:5000/` and double-click on the "failure reason" field
(in the right-hand test-run details pane) to edit it. Your change will
be POSTed to the web server, which will update the underlying
`failure-reason` file and the testrun and summary `index.html` files.

Note that if you don't want the interactive editing feature, you can
still view the html report locally (using a `file:///...` url), or you
can serve the report as static html files with a web server like apache.

You have to press "refresh" in your browser to see the failure-reason
change reflected in the summary table. This is because the test-run
details html is in a separate iframe, and I don't know how javascript in
the iframe can access the parent document.

`server` saves the user-given failure reason to `failure-reason.manual`,
so that re-running `report` won't overwrite the user's failure reason.
If `failure-reason.manual` exists it is displayed in the reports instead
of `failure-reason`.