extra/runner/run: If NoVideo, check Blackmagic video-capture hardware freeze
Sometimes the Blackmagic video-capture card fails and never delivers
video again until the PC is rebooted. If this happens we don't continue
running any tests (because they would all fail; restarting the PC
requires human intervention).

We can detect this by running a capture with `GST_DEBUG` enabled. If the
card is functional, then the message "VideoInputFrameArrived" is printed
to stderr *even if the Blackmagic card has no video on its input* (for
example if the set-top box is off). But if the card has failed then the
message is never printed.

For each of the following scenarios the GStreamer debug output looks
like:

* Working Blackmagic card, video source (the system under test) is on:

    activating push
    activatepush
    VideoInputFrameArrived: Frame received [<timecode>] - Valid Frame - Size: <n>

* Working Blackmagic card, video source is off:

    activating push
    activatepush
    VideoInputFrameArrived: Frame received - No input signal detected

  (Sometimes even when the video source is off, you can get
  "VideoInputFrameArrived: Frame received [<timecode>] - Valid Frame".
  Maybe this is only when the Blackmagic card has been configured to
  deliver the last seen frame when there is no video; I'm not sure.)

* Blackmagic card in use by another process (e.g. `stbt tv`):

    enable video input failed

  (and the GStreamer pipeline exits)

* Blackmagic card crashed (requiring a reboot of the PC):

    activating push
    activatepush

  (and the GStreamer pipeline just hangs after that output; eventually a
  stbt function like `wait_for_match` will timeout and raise `NoVideo`).

Note also that this writes to `failure-reason.manual`, because
`failure-reason` can be overwritten if you (manually) re-run `report`.
This Blackmagic failure can only be detected "online" (at the time the
test failed, on the same system that ran the test).
`failure-reason.manual` was introduced in commit 56aa1db9 to store
failure reasons entered manually by the user via the web interface, but
it is useful in this case too. Hopefully the "manual" name isn't too
confusing.

Thanks to Aiman Baharna <aiman.baharna@youview.com> for figuring out
how to detect Blackmagic failures.