extra/runner/run: Do Blackmagic failure detection before user recovery script
The user `recover` script might choose to power-cycle the
system-under-test, which isn't necessary if the failure was the fault
of the Blackmagic video-capture card.

`check_capture_hardware` now writes to `failure-reason` as well as
`failure-reason.manual` (even though `failure-reason` will be
overwritten the next time the user manually runs `report`) so that the
user `recover` script sees a failure reason of "Blackmagic card in use"
instead of "NoVideo".