extra/runner/server: Allow user to remove test results from report
This doesn't delete results altogether, it just hides them by renaming
the testrun folder with a leading ".".

We add the "Hide this result" link with javascript instead of adding it
to `templates/testrun.html`, so that the link only appears if a server
capable of handling it is running (i.e. the `server` script, which also
serves `edit-testrun.js`).

You'll still need to manually press "refresh" in the browser to see the
summary report without the deleted testrun. This is because the test-run
details html is in a separate iframe, and I don't know how javascript in
the iframe can access the parent document.

Note that we now need to call `report` so that it regenerates *only* the
summary html; call `report path/to/index.html`. I've also updated
`report.py` to provide a nicer error message if `report` is called in
this way with a path that doesn't contain testruns.

I'm testing this in the existing unit test rather than creating a
separate test, because the test setup is relatively complicated and it
takes a few secondns to run.