launcher: GUI wrapper for stb-tester commands
A simple GUI to launch stbt terminal commands. Users of stb-tester who
are not used to the command line should find it helpful. It could appeal
to users who refrain from using stb-tester because of the command line
"wizardy" it requires.

Mandatory command line arguments are specified through dialog boxes;
optional parameters are read from the stbt config file, or the default
values are used.

"config" menu item is different from stbt config: it brings up a
 simple text editor to modify the config file. ~/.config/stbt/stbt.conf
is opened by default, or the value of $STBT_CONFIG_FILE if it's in the
environment. If the default config file doesn't exist, the tool creates
it for the user.

Dialog boxes are created with Zenity. It's pre-installed with Gnome
and available in MacPorts. However, I was unable to make Zenity work on
a Mac.