extra/vm/setup.sh: Install `lsof`
Used by `extra/runner/run` since commit 313cf4fa. `lsof` is already
installed as a dependency of `ubuntu-standard`, but it is nice to list
it explicitly so that this list of packages can serve as documentation
to people who want to build their own stb-tester packages for Ubuntu
or other operating systems.

`lsof` isn't in `extra/stb-tester.spec` because that package doesn't
install the runner; for now the runner is only available from the source
repository.