extra/runner/run: Pass extra command-line arguments to test script
Commit bf393c1d taught `stbt run` to pass extra command-line arguments
to the test script. This current commit does the same for
`extra/runner/run`, the script for running multiple test scripts in
bulk. This allows you to run the same test script with different
arguments when you need to run multiple permutations of a test case.

To use this, place "--" on the command line between test scripts to run:

    runner/run test1.py arg1 arg2 -- test2.py ...

Without the "--", each argument is treated as a separate test, the same
way it worked before this commit. This makes it convenient to use bash's
brace expansion when you don't need to pass arguments to any of the test
scripts:

    runner/run very/long/path/to/{test1.py,test2.py,test3.py}

If you want to pass arguments to a single test script, place a "--" at
the end:

    runner/run test.py arg1 arg2 --

IMPLEMENTATION NOTES
--------------------

The implementation tries hard to preserve whitespace in arguments:

`read -a test` reads from standard input into the array variable `test`.

`IFS=$'\t' read ...` makes `read` split elements on tabs, not on any
whitespace (`$'\t'` is a string containing a single tab character).
Unfortunately you can't use `IFS=$'\0'` (split on the null character)
because bash internally uses C-style null-terminated strings. Tab is
a character that hopefully won't appear in genuine user arguments.

`"${test[@]}"` (and `"$@"` for the parameters to a function)
individually quotes each element in the array.

See:

http://www.gnu.org/software/bash/manual/html_node/Arrays.html
http://www.gnu.org/software/bash/manual/html_node/Bash-Builtins.html#index-read
http://www.gnu.org/software/bash/manual/html_node/ANSI_002dC-Quoting.html