stbt.press: Show key name in output video
This makes it a lot easier to understand what is happening when watching
a test run, and more importantly, when triaging the recorded video from
a failed test.

Note that `Display.push_buffer` has to iterate over a *copy* of the
`video_debug` list because it is removing elements from the (original)
list as it iterates over it^H^H the copy.

Note that we don't have to worry about the text overflowing the video
frame, either horizontally (for long text) or vertically (for many quick
`press`es). The OpenCV `putText` function just ignores areas outside of
the target image.