Motion detection reads default parameters from config with MotionParameters
As with the MatchParameters class introduced in commit 25d7f3a9, which
enabled global configuration of template matching parameters instead of
per match-call, this commit adds the same functionality for motion detect.

In the process, it exposes the `erode_passes` parameter for the motion
detect algorithm. Users should note that, whereas with templatematching
where increasing the number of erode passes increases makes it more likely to
get a false positive, increasing the number of erode passes in motion detect
is more likely to yield a false negative. (Zero is the "perfect" value for
both, however).

Both deployed and test `stbt.conf` files have been updated, along with
the selftests.

Note that with the introduction of MotionParameters, a bug was created with
`wait_for_motion` where if you specified motion_parameters but not a mask,
or specified them in the order, <parameters, mask>, then their values
were mixed up when passed to `detect_motion` because they were passed
positionally, not via keywords. Therefore, we `wait_for_motion` now
calls `detect_motion` explicitly with keywords.