Define ConfigurationError before a failed `get_config` can occur
If the stbt.py source code contains an invalid `get_config` e.g.
get_config("global", "invalid")
Then a `ConfigParser.Error` exception was caught so that we could raise
our own `ConfigurationError` exception.

However, *at import time* this hasn't be defined yet (it's during the
import that the exception is raised) and the Exception message
"NameError: global name 'ConfigurationError' is not defined"
is not useful to determining that the failure occurred because of an
invalid configuration parameters.

Note that this would never have leaked down to users because it is picked
up by `make check`, `make install`, etc. but it is useful for developers
to have the accurate Exception message.

Note also that I had to move `UITestError` as well because
`ConfigurationError` inherits from it.

Commiter's pre-merge message: (Is it necessary For ConfigurationError to
inherit from UITestError? It doesn't expand the functionality at all.)