wait_for_* functions call detect_* with explicit keywords arguments
With the introduction of MotionParameters, I found/created a bug with
`wait_for_motion` where if you specified motion_parameters but not a mask,
or specified them in the order, <parameters, mask>, then their values
were mixed up when passed to `detect_motion` because they were passed
positionally, not via keywords.

For completedness and also good practice, I applied the same change to the
call to `detect_match` inside `wait_for_match`.