MotionParameters.noise_threshold renamed MotionParameters.motion_threshold
This is to avoid confusion when users uses the legacy `noise_threshold`
parameter which raises a Warning telling the user it is deprecated and
to use `MotionParameters.noise_threshold` instead. It's much clearer to
just rename the parameter entirely.

The fact that the `noise/motion_threshold` is useful if you have a noisy
video to analyse is somewhat by-the-by; ultimately it controls how much
*motion* (difference between consecutive frames) is needed to pass the
motion detect test.

Note: When changing the name of the selftest functions I introduced an
underscore between "motion" and "threshold" so that it gets found when
doing a `(git) grep`.