Customise detect_motion visualisation colour with MotionParameters
Users can now customise the detect_motion highlight visualisation colour
using MotionParameters and stbt.conf. A useful application of this is
to colour separate calls to `wait_for_motion` in a single test so that
the highlights are different colours, so when reviewing the test video
(when using --save-video) it is easier to distinguish the calls. This could
also be useful to users who have implemented a `wait_for_no_motion`
wrapper for `detect_motion` and colour motion when in is wanted
(`wait_for_motion`) and motion when it is not wanted (`wait_for_no_motion`)
different colours.

Note that due to use of OpenCV in stb-tester's framework, the colours are
specified Blue-Green-Red (BGR).